Use relative URL for dist location in examples
Because GitHub pages are not served from a root URL on anything other
than a user or organisation repo, the relative URL to the parent
directory is nessesary.

This still permits sites that are served from a root URL as a source URL
starting with `../dist/` on a root webpage with resolve to `/dist/`. As
such this does not break the local dev command `npm run dev`.

This also has the additional advantage that if a developer simply clones
the repo to view examples, then they do not need to serve the files from
a local server, they can simply access them through `file://` protocol.